backend/requirements.txt
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
Babel==2.14.0
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configparser==6.0.0
coverage==7.4.3
distro==1.9.0
dnspython==2.6.1
docutils==0.20.1
exceptiongroup==1.2.0
fastapi==0.109.2
flower==2.0.1
freezegun==1.4.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
humanize==4.9.0
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.3
kombu==5.3.5
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
mongoengine==0.27.0
mypy-extensions==1.0.0
openai==1.12.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
pytest==8.0.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==1.1.0
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
starkbank-ecdsa==2.2.0
starlette==0.36.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
